requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
idna==3.8
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==2.1.0
pandas==2.2.2
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
